requirements.txt
attrs==17.2.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.1.2
enum34==1.1.6
future==0.16.0
futures==3.1.1
greenlet==0.4.12
ipython==5.4.1
ipython-genutils==0.2.0
jedi==0.10.2
msgpack-python==0.4.8
neovim==0.1.13
pathlib2==2.3.0
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyflakes==1.6.0
Pygments==2.2.0
pymongo==3.5.0
python-dateutil==2.6.1
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
trollius==2.1
wcwidth==0.1.7
